# docs/api_contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScoutBoy API",
    "description": "ScoutBoy \u2014 FUT.gg-style real-life player discovery. MVP scope: U23 attackers and midfielders in Europe. Every score/badge/value is explainable.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/players": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Search Players",
        "operationId": "search_players_api_players_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "age_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Age Min"
            }
          },
          {
            "name": "age_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Age Max"
            }
          },
          {
            "name": "position_group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Position Group"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "League"
            }
          },
          {
            "name": "club",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Club"
            }
          },
          {
            "name": "nationality",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Nationality"
            }
          },
          {
            "name": "min_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Minutes"
            }
          },
          {
            "name": "rolefit_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rolefit Min"
            }
          },
          {
            "name": "rolefit_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rolefit Max"
            }
          },
          {
            "name": "playstyle",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Playstyle"
            }
          },
          {
            "name": "value_min",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Value Min"
            }
          },
          {
            "name": "value_max",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Value Max"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "rolefit_desc",
              "title": "Sort"
            }
          },
          {
            "name": "universe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "'mvp' (default) filters to the U23 EU att/mid universe; 'all' includes everyone",
              "default": "mvp",
              "title": "Universe"
            },
            "description": "'mvp' (default) filters to the U23 EU att/mid universe; 'all' includes everyone"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_PlayerSearchCard_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/{player_id}": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Player",
        "operationId": "get_player_api_players__player_id__get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/{player_id}/ratings": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Player Ratings",
        "operationId": "get_player_ratings_api_players__player_id__ratings_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleRatingDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/{player_id}/playstyles": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Player Playstyles",
        "operationId": "get_player_playstyles_api_players__player_id__playstyles_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerPlaystylesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/{player_id}/market": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Player Market",
        "operationId": "get_player_market_api_players__player_id__market_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketPanel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/{player_id}/similar": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Similar Players",
        "operationId": "get_similar_players_api_players__player_id__similar_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimilarResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/ingest": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Trigger Ingest",
        "operationId": "trigger_ingest_api_admin_ingest_post",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "sample",
              "title": "Source"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/recompute-ratings": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Trigger Recompute",
        "operationId": "trigger_recompute_api_admin_recompute_ratings_post",
        "parameters": [
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecomputeResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/rating-runs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Rating Runs",
        "operationId": "rating_runs_api_admin_rating_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RatingRunSummary"
                  },
                  "title": "Response Rating Runs Api Admin Rating Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestResult": {
        "properties": {
          "run_id": {
            "type": "integer",
            "title": "Run Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "players": {
            "type": "integer",
            "title": "Players"
          },
          "metrics": {
            "type": "integer",
            "title": "Metrics"
          },
          "quality_errors": {
            "type": "integer",
            "title": "Quality Errors"
          },
          "quality_warnings": {
            "type": "integer",
            "title": "Quality Warnings"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "source",
          "players",
          "metrics",
          "quality_errors",
          "quality_warnings"
        ],
        "title": "IngestResult"
      },
      "MarketPanel": {
        "properties": {
          "public_value_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Value Eur"
          },
          "model_value_low_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Value Low Eur"
          },
          "model_value_high_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Value High Eur"
          },
          "expected_asking_low_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Asking Low Eur"
          },
          "expected_asking_high_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Asking High Eur"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence",
            "default": "unknown"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": "unknown"
          },
          "manual_review_required": {
            "type": "boolean",
            "title": "Manual Review Required",
            "default": false
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "explanation": {
            "additionalProperties": true,
            "type": "object",
            "title": "Explanation",
            "default": {}
          }
        },
        "type": "object",
        "title": "MarketPanel"
      },
...
      "Paginated_PlayerSearchCard_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlayerSearchCard"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "Paginated[PlayerSearchCard]"
      },
      "PlayerCardResponse": {
        "properties": {
          "identity": {
            "$ref": "#/components/schemas/PlayerIdentity"
          },
          "season": {
            "type": "string",
            "title": "Season"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "best_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Role"
          },
          "face_stats": {
            "items": {
              "$ref": "#/components/schemas/FaceStat"
            },
            "type": "array",
            "title": "Face Stats",
            "default": []
          },
          "substats": {
            "items": {
              "$ref": "#/components/schemas/SubStat"
            },
            "type": "array",
            "title": "Substats",
            "default": []
          },
          "role_ratings": {
            "items": {
              "$ref": "#/components/schemas/RoleRatingSummary"
            },
            "type": "array",
            "title": "Role Ratings",
            "default": []
          },
          "playstyles": {
            "items": {
              "$ref": "#/components/schemas/PlaystyleBadge"
            },
            "type": "array",
            "title": "Playstyles",
            "default": []
          },
          "concerns": {
            "items": {
              "$ref": "#/components/schemas/PlaystyleBadge"
            },
            "type": "array",
            "title": "Concerns",
            "default": []
          },
          "market": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarketPanel"
              },
              {
                "type": "null"
              }
            ]
          },
          "strengths": {
            "items": {
              "$ref": "#/components/schemas/StrengthConcern"
            },
            "type": "array",
            "title": "Strengths",
            "default": []
          },
          "concerns_text": {
            "items": {
              "$ref": "#/components/schemas/StrengthConcern"
            },
            "type": "array",
            "title": "Concerns Text",
            "default": []
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextPanel"
              },
              {
                "type": "null"
              }
            ]
          },
          "data_sources": {
            "items": {
              "$ref": "#/components/schemas/DataSource"
            },
            "type": "array",
            "title": "Data Sources",
            "default": []
          },
          "last_updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          },
          "rating_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating Version"
          }
        },
        "type": "object",
        "required": [
          "identity",
          "season",
          "confidence"
        ],
        "title": "PlayerCardResponse"
      },
...
      "PlayerPlaystylesResponse": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "season": {
            "type": "string",
            "title": "Season"
          },
          "playstyles": {
            "items": {
              "$ref": "#/components/schemas/PlaystyleBadge"
            },
            "type": "array",
            "title": "Playstyles",
            "default": []
          },
          "concerns": {
            "items": {
              "$ref": "#/components/schemas/PlaystyleBadge"
            },
            "type": "array",
            "title": "Concerns",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "season"
        ],
        "title": "PlayerPlaystylesResponse"
      },
...
      "RecomputeResult": {
        "properties": {
          "run_id": {
            "type": "integer",
            "title": "Run Id"
          },
          "affected": {
            "type": "integer",
            "title": "Affected"
          },
          "seasons": {
            "type": "integer",
            "title": "Seasons"
          },
          "outlier_check": {
            "additionalProperties": true,
            "type": "object",
            "title": "Outlier Check",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "affected",
          "seasons"
        ],
        "title": "RecomputeResult"
      },
...
      "RoleRatingDetail": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "season": {
            "type": "string",
            "title": "Season"
          },
          "rating_version": {
            "type": "string",
            "title": "Rating Version"
          },
          "ratings": {
            "items": {
              "$ref": "#/components/schemas/RoleRatingSummary"
            },
            "type": "array",
            "title": "Ratings",
            "default": []
          },
          "audits": {
            "items": {
              "$ref": "#/components/schemas/AuditBreakdown"
            },
            "type": "array",
            "title": "Audits",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "season",
          "rating_version"
        ],
        "title": "RoleRatingDetail"
      },
...
      "SimilarResponse": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "season": {
            "type": "string",
            "title": "Season"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/SimilarGroup"
            },
            "type": "array",
            "title": "Groups",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "season"
        ],
        "title": "SimilarResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}